Add ~veracity to version number in order to distinguish it from candor builds.
The epoch number "1" is also added, because adding the ~veracity tag made the
version smaller than any previous numbers without the tag. The number "2" will
be used for candor versions.

This resolves naming conflict between Veracity and Candor builds.
This naming conflict led to a "clogged" queue when pushing to the package repo.